move: GlobalErrorCode, GenericResponse, BaseEntity 패키지화
